requirements.txt
PyPDF2==3.0.1
Pillow==10.0.0
pdf2image==1.16.3
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.66.1 